Navigating to match while keeping stuff in the modules.
